refactor(student profile): change folders and files names
